Desarrollo_ServidorDjangoCompleto/requirements.txt
pandas
Django
requests
djangorestframework
mysqlclient